+++ Samples/iOS-Swift/iOS-Swift/Profiling/ProfilingViewController.swift
@@ -1,3 +1,4 @@
+import Sentry
 import UIKit
 
 class ProfilingViewController: UIViewController, UITextFieldDelegate {
@@ -103,16 +104,24 @@ class ProfilingViewController: UIViewController, UITextFieldDelegate {
         }
     }
     
+    @IBAction func startContinuousProfiler(_ sender: Any) {
+        SentrySDK.startProfiler()
+    }
+    
+    @IBAction func stopContinuousProfiler(_ sender: Any) {
+        SentrySDK.stopProfiler()
+    }
+    
     @IBAction func viewLastProfile(_ sender: Any) {
         profilingUITestDataMarshalingTextField.text = "<fetching...>"
-        withProfile(fileName: "profile") { file in
+        withProfile(first: false) { file in
             handleContents(file: file)
         }
     }
     
-    @IBAction func viewLaunchProfile(_ sender: Any) {
+    @IBAction func viewFirstContinuousProfileChunk(_ sender: Any) {
         profilingUITestDataMarshalingTextField.text = "<fetching...>"
-        withProfile(fileName: "profile") { file in
+        withProfile(first: true) { file in
             handleContents(file: file)
         }
     }
@@ -123,16 +132,25 @@ class ProfilingViewController: UIViewController, UITextFieldDelegate {
         textField.resignFirstResponder()
     }
     
-    private func withProfile(fileName: String, block: (URL?) -> Void) {
+    private func withProfile(first: Bool, block: (URL?) -> Void) {
         let appSupportDirectory = NSSearchPathForDirectoriesInDomains(.applicationSupportDirectory, .userDomainMask, true).first!
+        let fm = FileManager.default
+        let dir = "\(appSupportDirectory)/io.sentry/profiles"
+        let count = try! fm.contentsOfDirectory(atPath: dir).count
+        //swiftlint:disable empty_count
+        guard first || count > 0 else {
+            //swiftlint:enable empty_count
+            profilingUITestDataMarshalingTextField.text = "<missing>"
+            return
+        }
+        let fileName = "profile\(first ? 0 : count - 1)"
+        let fullPath = "\(dir)/\(fileName)"
         
-        let fullPath = "\(appSupportDirectory)/io.sentry/\(fileName)"
-        
-        if FileManager.default.fileExists(atPath: fullPath) {
+        if fm.fileExists(atPath: fullPath) {
             let url = NSURL.fileURL(withPath: fullPath)
             block(url)
             do {
-                try FileManager.default.removeItem(atPath: fullPath)
+                try fm.removeItem(atPath: fullPath)
             } catch {
                 SentrySDK.capture(error: error)
             }