+++ Samples/iOS-Swift/iOS-Swift/Tools/iOS-Swift-Bridging-Header.h
@@ -3,3 +3,4 @@
 #import "SentryUIApplication.h"
 #import <Sentry/PrivateSentrySDKOnly.h>
 #import <Sentry/SentryOptions+Private.h>
+#import <Sentry/SentrySDK+Private.h>